backend-app-api: move global loggers and winston helper back to backend-common
